adaption for pypi
