+++ realesrgan/__init__.py
@@ -2,4 +2,5 @@
 from .archs import *
 from .data import *
 from .models import *
+from .utils import *
 from .version import __gitsha__, __version__
